spoof-detection-app/requirements.txt
#  pip install -r requirements.txt
Flask>=2.2.0
flask-cors>=3.0.10
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.24.0
opencv-python>=4.7.0
pillow>=9.4.0
mediapipe>=0.9.0
transformers>=4.30.0
scikit-learn>=1.3.2
ipython>=8.0.0